requirements.txt
# Pi dependencies (hardware)
gpiozero
luma.lcd
sensirion-i2c-driver
sensirion-i2c-scd

# PC preview only
pygame
pillow
